move system.prop settings to p1.mk PRODUCT_PROPERTY_OVERRIDES
This is a workaround for a limitation of the android build system because
system.prop is only referenced via $(TARGET_DEVICE_DIR)/system.prop.

To avoid the duplication of p1/system.prop in the p1{l,n} repositories
it is necessary to move these settings to p1.mk and get rid of system.prop
entirely. As a side effect all property settings are now in one place.

Change-Id: I3b6786c63f677491a830045eee70fadf4afa9ef5