Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

nothing wrong with cleanup

however, most of these settings that you're shuffling around are where they are for the sake of continuity with other devices

for example, kernel config settings, releasetools, etc belong in BoardConfig.mk as every other device in cm has them there. If someone were to look at our device trees as references it adds to the confusion if they see something missing, or have to go hunting for it, as opposed to other devices.

Patch-set: 1