Update patch set 1
Patch Set 1:

The binaries from system/bin are not static. I am not sure is it important, but binaries from utilities are. There are static versions of needed utils that could be added to device_base.mk and that are copied during building to utilities directory by standard build procedure. I could upload patch.

Patch-set: 1